Fix lint
